PAVEPO/audio_service/config.py
from pydantic_settings import BaseSettings

class Settings(BaseSettings):
    DATABASE_URL: str = "postgresql+asyncpg://postgres:postgres@db:5432/audio_service"
    SECRET_KEY: str = "secret"
    ALGORITHM: str = "HS256"
    ACCESS_TOKEN_EXPIRE_MINUTES: int = 30
    YANDEX_CLIENT_ID: str = ''
    YANDEX_CLIENT_SECRET: str = ''
    YANDEX_REDIRECT_URI: str = "http://localhost:8000/auth/yandex/callback"
    AUDIO_STORAGE_PATH: str = "./audio_files"

    class Config:
        env_file = ".env"

settings = Settings()