PAVEPO/audio_service/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-jose==3.3.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6
passlib==1.7.4
alembic==1.13.1